[refactor] Change function arg type

+++ srcs/ft_substr.c
@@ -6,13 +6,13 @@
 /*   By: ttomori <ttomori@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/06 22:26:02 by ttomori           #+#    #+#             */
-/*   Updated: 2022/01/20 00:07:19 by ttomori          ###   ########.fr       */
+/*   Updated: 2022/01/20 00:10:42 by ttomori          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_substr(const char *s, unsigned int start, size_t maxlen)
+char	*ft_substr(const char *s, size_t start, size_t maxlen)
 {
 	char	*p;
 	size_t	len;
@@ -21,7 +21,7 @@ char	*ft_substr(const char *s, unsigned int start, size_t maxlen)
 	if (s == NULL)
 		return (NULL);
 	len = ft_strlen(s);
-	if (len <= (size_t)start)
+	if (len <= start)
 		return (ft_strdup(""));
 	size = ft_strlen(s + start);
 	if (maxlen < size)
